requirements.txt
# Conteúdo sugerido para requirements.txt (na raiz do projeto)
pyyaml
crewai
pydantic==2.5.3
openai
python-dotenv
pypdf # Para extração de PDF
chromadb # Exemplo de Vector DB para o Memory Agent